Resolve comments and issues

+++ src/database/migrations/2020_12_15_122241_add_posted_to_favorites_table.php
@@ -25,8 +25,12 @@ class AddPostedToFavoritesTable extends Migration
      */
     public function down()
     {
-        Schema::table('favorites', function (Blueprint $table) {
-            $table->dropColumn('posted');
-        });
+        if (Schema::hasColumn('favorites', 'posted')) {
+
+            Schema::table('favorites', function (Blueprint $table) {
+
+                $table->dropColumn('posted');
+            });
+        }
     }
 }
